+++ GccSharp/GccSharp.ConsoleApp/Program.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using CommandLine;
 using GccSharp.ConsoleApp.Arguments;
+using GccSharp.ConsoleApp.Confirmation;
+using GccSharp.ConsoleApp.Processors;
 
 namespace GccSharp.ConsoleApp
 {
@@ -56,57 +58,6 @@ namespace GccSharp.ConsoleApp
             return false;
         }
 
-        private static bool ConsoleConfirmation(Activity activity)
-        {            
-            Console.WriteLine(activity);
-            Console.WriteLine("Do you wish to process this entry? (y/n)");            
-            return Console.ReadKey(true).Key == ConsoleKey.Y;
-        }
-
-        private static bool WebProcessor(Activity activity)
-        {
-            var successful = true;            
-            Console.WriteLine("\r\n\tEntering Web Processor \r\n ");
-            try
-            {
-                var clientEmail = GetClientEmail();
-                var clientPassword = GetClientPassword();
-                using (var client = new Client())
-                {
-                    client.Login(clientEmail, clientPassword);
-                    client.Submit(activity);
-                    client.Logout();
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine("\r\nError: \r\n" + ex);
-                successful = false;
-            }
-            Console.WriteLine("\r\n\tExiting Web Processor \r\n ");
-            return successful;
-        }
-
-        private static string GetClientEmail()
-        {
-            var email = Configuration.ClientEmail;
-            if (!string.IsNullOrWhiteSpace(email)) return email;
-
-            Console.WriteLine("Could not find email address.");
-            Console.Write("Email: ");
-            return Console.ReadLine();            
-        }
-
-        private static string GetClientPassword()
-        {
-            var password = Configuration.ClientPassword;
-            if (!string.IsNullOrWhiteSpace(password)) return password;
-
-            Console.WriteLine("Could not find password.");
-            Console.Write("Password: ");
-            return Console.ReadLine();            
-        }
-
         private static Activity ExtractActivity(string verbName, object verbSubOptions)
         {
             Activity activity = null;
@@ -159,11 +110,12 @@ namespace GccSharp.ConsoleApp
         {
             if (Processor == null)
             {
-                Processor = WebProcessor;
+                Processor = WebProcessor.Processor;
             }
             if (Confirmation == null)
             {
-                Confirmation = ConsoleConfirmation;
+                Confirmation = ConsoleConfirmation.Confirmation;
+                //Confirmation = MessageBoxConfirmation.Confirmation;
             }
         }
